add code to do fasta.gz and move some file to the Accessory folder, feng 1/8/2019

+++ Accessory/FastaHandler.hpp
@@ -12,7 +12,9 @@ using namespace std;
 
 //return total number of sequeces read in.
 //toUpper =true : convert the character to upper case. false: keep the original letter
-unsigned int ReadFasta(const string& _fname, vector<SequenceString>& _seqStrVec, bool toUpper=false);
+//now this can handle automatically either fasta or gzip'ed fasta 
+//
+size_t ReadFasta(const string& _fname, vector<SequenceString>& _seqStrVec, bool toUpper=false);
 
 
 void WriteFasta(const string& _fname, vector<SequenceString>& _seqStrVec, const unsigned int& _width=50, ios_base::openmode mode=ios_base::out);
