+++ Accessory/FastqHandler.hpp
@@ -3,9 +3,9 @@
 #include <string>
 #include <vector>
 #include <iostream>
-
+#include <string.h>
 #include "SequenceString.hpp"
-#include "Accessory/GzTools.hpp"
+#include "GzTools.hpp"
 #include "FASTQ.hpp"
 
 //here, we won't define a class, instead we simply write up a few functions to
@@ -17,7 +17,8 @@ using namespace std;
 //
 //the file could be gz'ed or regular fastq. then the sequence data are read into a vector of 
 //fastq objects. 
-unsigned int ReadFastq(const string& _fname, vector<Fastq>& _seqStrVec, bool toUpper=false);
+//return string::npos upon error 
+size_t ReadFastq(const string& _fname, vector<Fastq>& _seqStrVec, bool toUpper=false);
 
 
 void WriteFastq(const string& _fname, vector<Fastq>& _seqStrVec,  ios_base::openmode mode=ios_base::out);
